requirements.txt
fastapi[standard]
pymongo
pyjwt[crypto]
pathlib
bcrypt
python-multipart
pydantic_settings
motor